# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Car Rental API",
    "description": "API for managing car rentals"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "schemes": ["http", "https"],
  "produces": ["application/json"],
  "consumes": ["application/json"],
  "paths": {
    "/data": {
      "get": {
        "summary": "Get car data",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Car"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create a new order",
        "parameters": [
          {
            "in": "body",
            "name": "order",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order created successfully",
            "schema": {
              "$ref": "#/definitions/OrderOutput"
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "User login",
        "parameters": [
          {
            "in": "body",
            "name": "credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginCredentials"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "401": {
            "description": "Invalid username or password"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "User registration",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRegistration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registration successful"
          },
          "400": {
            "description": "Invalid password format"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/available-cars": {
      "post": {
        "summary": "Get available cars",
        "parameters": [
          {
            "in": "body",
            "name": "dates",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AvailableCarsInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Car"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "summary": "Refresh access token",
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "schema": {
              "$ref": "#/definitions/TokenRefreshResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Get user data",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "406": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "User logout",
        "responses": {
          "204": {
            "description": "Logout successful"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/getOrders": {
      "get": {
        "summary": "Get user orders",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order"
              }
            }
          },
          "401": {
            "description": "Refresh token not provided"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/add-to-wishlist": {
      "post": {
        "summary": "Add car to wishlist",
        "parameters": [
          {
            "in": "body",
            "name": "car",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WishlistInput"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Car added to wishlist successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/getFavorites": {
      "get": {
        "summary": "Get user favorites",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Refresh token not provided"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/getCarID": {
      "get": {
        "summary": "Get car ID",
        "parameters": [
          {
            "in": "query",
            "name": "carName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/CarIDResponse"
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/removeFromFavorites": {
      "delete": {
        "summary": "Remove car from favorites",
        "parameters": [
          {
            "in": "query",
            "name": "carID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Car removed from favorites successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/searchCars": {
      "get": {
        "summary": "Search for cars",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Car"
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "definitions": {
    "Car": {
      "type": "object"
    },
    "OrderInput": {
      "type": "object"
    },
    "OrderOutput": {
      "type": "object"
    },
    "LoginCredentials": {
      "type": "object"
    },
    "LoginResponse": {
      "type": "object"
    },
    "UserRegistration": {
      "type": "object"
    },
    "AvailableCarsInput": {
      "type": "object"
    },
    "TokenRefreshResponse": {
      "type": "object"
    },
    "UserResponse": {
      "type": "object"
    },
    "Order": {
      "type": "object"
    },
    "WishlistInput": {
      "type": "object"
    },
    "CarIDResponse": {
      "type": "object"
    }
  }
}